requirements.txt
streamlit==1.31.1
numpy==1.26.2
pandas==2.1.3
plotly==5.18.0
yfinance==0.2.31
scipy==1.11.3
scikit-learn==1.3.2
python-dotenv==1.0.0
google-generativeai==0.3.2
PyQt5==5.15.10
matplotlib==3.8.2